+++ unified-runtime/source/adapters/hip/queue.cpp
@@ -28,89 +28,17 @@ void ur_queue_handle_t_::transferStreamWaitForBarrierIfNeeded(
   }
 }
 
-hipStream_t ur_queue_handle_t_::getNextComputeStream(uint32_t *StreamToken) {
-  if (getThreadLocalStream() != hipStream_t{0})
-    return getThreadLocalStream();
-  uint32_t Stream_i;
-  uint32_t Token;
-  while (true) {
-    if (NumComputeStreams < ComputeStreams.size()) {
-      // the check above is for performance - so as not to lock mutex every time
-      std::lock_guard<std::mutex> guard(ComputeStreamMutex);
-      // The second check is done after mutex is locked so other threads can not
-      // change NumComputeStreams after that
-      if (NumComputeStreams < ComputeStreams.size()) {
-        UR_CHECK_ERROR(hipStreamCreateWithPriority(
-            &ComputeStreams[NumComputeStreams++], Flags, Priority));
-      }
-    }
-    Token = ComputeStreamIdx++;
-    Stream_i = Token % ComputeStreams.size();
-    // if a stream has been reused before it was next selected round-robin
-    // fashion, we want to delay its next use and instead select another one
-    // that is more likely to have completed all the enqueued work.
-    if (DelayCompute[Stream_i]) {
-      DelayCompute[Stream_i] = false;
-    } else {
-      break;
-    }
-  }
-  if (StreamToken) {
-    *StreamToken = Token;
-  }
-  hipStream_t Res = ComputeStreams[Stream_i];
-  computeStreamWaitForBarrierIfNeeded(Res, Stream_i);
-  return Res;
+ur_queue_handle_t ur_queue_handle_t_::getEventQueue(const ur_event_handle_t e) {
+  return e->getQueue();
 }
 
-hipStream_t ur_queue_handle_t_::getNextComputeStream(
-    uint32_t NumEventsInWaitList, const ur_event_handle_t *EventWaitList,
-    ur_stream_guard &Guard, uint32_t *StreamToken) {
-  if (getThreadLocalStream() != hipStream_t{0})
-    return getThreadLocalStream();
-  for (uint32_t i = 0; i < NumEventsInWaitList; i++) {
-    uint32_t Token = EventWaitList[i]->getComputeStreamToken();
-    if (EventWaitList[i]->getQueue() == this && canReuseStream(Token)) {
-      std::unique_lock<std::mutex> ComputeSyncGuard(ComputeStreamSyncMutex);
-      // redo the check after lock to avoid data races on
-      // LastSyncComputeStreams
-      if (canReuseStream(Token)) {
-        uint32_t Stream_i = Token % DelayCompute.size();
-        DelayCompute[Stream_i] = true;
-        if (StreamToken) {
-          *StreamToken = Token;
-        }
-        Guard = ur_stream_guard{std::move(ComputeSyncGuard)};
-        hipStream_t Res = EventWaitList[i]->getStream();
-        computeStreamWaitForBarrierIfNeeded(Res, Stream_i);
-        return Res;
-      }
-    }
-  }
-  Guard = {};
-  return getNextComputeStream(StreamToken);
+uint32_t
+ur_queue_handle_t_::getEventComputeStreamToken(const ur_event_handle_t e) {
+  return e->getComputeStreamToken();
 }
 
-hipStream_t ur_queue_handle_t_::getNextTransferStream() {
-  if (getThreadLocalStream() != hipStream_t{0})
-    return getThreadLocalStream();
-  if (TransferStreams.empty()) { // for example in in-order queue
-    return getNextComputeStream();
-  }
-  if (NumTransferStreams < TransferStreams.size()) {
-    // the check above is for performance - so as not to lock mutex every time
-    std::lock_guard<std::mutex> Guard(TransferStreamMutex);
-    // The second check is done after mutex is locked so other threads can not
-    // change NumTransferStreams after that
-    if (NumTransferStreams < TransferStreams.size()) {
-      UR_CHECK_ERROR(hipStreamCreateWithPriority(
-          &TransferStreams[NumTransferStreams++], Flags, Priority));
-    }
-  }
-  uint32_t Stream_i = TransferStreamIdx++ % TransferStreams.size();
-  hipStream_t Res = TransferStreams[Stream_i];
-  transferStreamWaitForBarrierIfNeeded(Res, Stream_i);
-  return Res;
+hipStream_t ur_queue_handle_t_::getEventStream(const ur_event_handle_t e) {
+  return e->getStream();
 }
 
 UR_APIEXPORT ur_result_t UR_APICALL
@@ -246,18 +174,13 @@ UR_APIEXPORT ur_result_t UR_APICALL urQueueRelease(ur_queue_handle_t hQueue) {
 }
 
 UR_APIEXPORT ur_result_t UR_APICALL urQueueFinish(ur_queue_handle_t hQueue) {
-  // set default result to a negative result (avoid false-positve tests)
-  ur_result_t Result = UR_RESULT_ERROR_OUT_OF_RESOURCES;
+  ur_result_t Result = UR_RESULT_SUCCESS;
 
   try {
-
     ScopedDevice Active(hQueue->getDevice());
 
-    hQueue->syncStreams<true>([&Result](hipStream_t S) {
-      UR_CHECK_ERROR(hipStreamSynchronize(S));
-      Result = UR_RESULT_SUCCESS;
-    });
-
+    hQueue->syncStreams<true>(
+        [&Result](hipStream_t S) { UR_CHECK_ERROR(hipStreamSynchronize(S)); });
   } catch (ur_result_t Err) {
     Result = Err;
   } catch (...) {
